+++ CHANGELOG.md
@@ -6,6 +6,24 @@ All notable changes to magpylib are documented here.
 
 # Releases
 
+## [2.3.0b] - 2020-01-17
+
+### Changed
+- Improved performance of getB for diametral magnetized Cylinders by 20%.
+- GetB of Line current now uses vectorized code which leads to massive performance enhancement.
+- **IMPORTANT:** position arguments of `getBv` functions have been flipped! First comes the source position POSm THEN the observer position POSo!
+- - getB(pos) now takes single AND vector position arguments. If a vector is handed to getB it will automatically execute vectorized code from the vector module.
+
+### Added
+- completed the library vector functionality adding magnet Cylinder, moment Dipole, current Circular and Line. This includes adding several private vectorized functions (e.g. ellipticV) to mathLib_vector, adding respective tests and docu examples.
+
+---
+
+## [2.2.0b] - 2019-12-27
+- unreleased version
+
+---
+
 ## [2.1.0b] - 2019-12-06
 
 ### Added
@@ -23,18 +41,20 @@ All notable changes to magpylib are documented here.
 - Restructuring
   - displaySystem is now a top-level function, not a Collection method anymore.
   - getBsweep and multiprocessing options have been completely removed, this functionality
-    should be overtaken by the new vector functionality which uses numpy native vectorized 
+    should be overtaken by the new vector functionality which uses the numpy native vectorized 
     code paradigm. If mkl library is set (test by numpy.show_config()) numpy will also 
     automatically use multiporcessing. Code parallelization at magpylib level should be done
     by hand.
 - Docstrings are adjusted to work better with intellisense. (Problems with *.rst code)
 - public rotatePosition() is now called angleAxisRotation(), former private angleAxisRotation
     is now called angleAxisRotation_priv().
-- Major rework of the documentation and exmples.
+- Major rework of the documentation and examples.
 
 ### Added
 - Performance computation trough vector functionality included in new top-level subpackge "vector"
 - Vectorized versions of math functions added to "math" subpackage
+
+---
 
 ## [1.2.1b0] - 2019-07-31
 ### Changed